Filled files of migrations and models
