+++ app/Models/Dish.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Dish extends Model
+{
+    protected $fillable = [
+        'name',
+        'description',
+        'category_id',
+        'restaurant_id'
+    ];
+}
