+++ app/Models/DishCategory.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
 class DishCategory extends Model
 {
-    //
+    protected $fillable = [
+        'name'
+    ];
 }
