+++ app/Models/Event.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Event extends Model
+{
+    protected $fillable = [
+        'name',
+        'start_date',
+        'end_date',
+        'address',
+        'description',
+        'category_id'
+    ];
+}
