+++ app/Models/EventCategory.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
 class EventCategory extends Model
 {
-    //
+    protected $fillable = [
+        'name'
+    ];
 }
