+++ app/Models/Ingredient.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
 class Ingredient extends Model
 {
-    //
+    protected $fillable = [
+        'name'
+    ];
 }
