+++ app/Models/Recipe.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Recipe extends Model
+{
+    protected $fillable = [
+        'name',
+        'description',
+        'category_id',
+        'preparation_time',
+        'difficulty',
+        'products'
+    ];
+}
