+++ app/Models/Restaurant.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Restaurant extends Model
+{
+    protected $fillable = [
+        'name',
+        'address',
+        'description',
+        'business_hours',
+        'site_url'
+    ];
+}
