+++ app/Models/Shop.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Shop extends Model
+{
+    protected $fillable = [
+        'name',
+        'description',
+        'site_url'
+    ];
+}
